pyproject.toml
[project]
name = "qa"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "docx2txt (>=0.9,<0.10)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "huggingface-hub (>=0.32.4,<0.33.0)",
    "llama-cpp-python (>=0.3.9,<0.4.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "vllm (>=0.9.0.1,<0.10.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "dotenv (>=0.9.9,<0.10.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
